cli/test-paystack-check/backend/requirements.txt
# =============================================================================
# Test Paystack Check - Python Dependencies
# =============================================================================

# Core
fastapi==0.135.3
uvicorn[standard]==0.44.0
pydantic[email]==2.12.5
pydantic-settings==2.13.1

# Database - Tortoise ORM
tortoise-orm[asyncpg,ipython]==1.1.7
asyncpg==0.31.0

# Redis
redis==7.4.0

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
httpx==0.28.1
itsdangerous==2.2.0

# Rate Limiting
slowapi==0.1.9

# Utils
python-dotenv==1.0.1
python-multipart==0.0.20
aiofiles==24.1.0
jinja2==3.1.6

# Email - Resend
resend==2.19.0

# Monitoring (optional)
sentry-sdk[fastapi]==2.47.0

# Testing
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==7.1.0
aiosqlite==0.22.1
